patch sonos_hdmi_cec_tx_corruption.patch
This patch fixes a TX corruption issue related to the MediaTek CEC driver.

The driver resets the TX function on every call to mtk_cec_send_msg()
regardless of it was a new message or a continuation of an old message.
If the latter, the device would discard anything pending to transmit and
replace it with the new data.  This usually results in a message which
contains only the the first byte followed by the last 4 bytes of the
intended transmission.

The fix is to only perform the reset on a new message transmission only.

diff -aurp -x .unpacked -x '*.orig' linux-4.4.24-mtk-orig/linux-4.4.24-mtk/drivers/gpu/drm/mediatek/mtk_cec.c linux-4.4.24-mtk/drivers/gpu/drm/mediatek/mtk_cec.c