patch sonos_hdmi_cec_nack_errors.patch
The MTK CEC driver has three types of "errors" which are handled together
in the transmit event handler: LOWB (low buffer), UN (underflow), and FAIL.
LOWB and UN are transmit failures that occur at the hardware/driver level
and producing a dev_err() message is probably the correct thing to do.

FAIL however describes a transmit failure at the physical layer due to
network activity.  It either indicates a packet collision which prevented
transmission or that the transmit succeeded but was not ACKed by the
destination device(s).  It is this second case we are encountering.

When a CEC device is brought up it must attempt to obtain the desired
Logical Address (LA) which must be unique to the network.  To determine
availability it "pings" the LA four times, waiting for an acknowledge
each time.  If there are no ACK then it is assumed the LA is available.
The MTK CEC driver is reporting an error for the 4 pings not being ACKed,
which is perfectly normal and not an error.

The errors also show up whenever any other unicast message goes unACKed
or a broadcast message gets explicitly NACKed.  This behavior occurs a
lot during topology scans as many destination LAs probed are not allocated
and go unACKed, thus generating more errors.

diff -aurp -x .unpacked -x '*.orig' linux-4.4.24-mtk-orig/linux-4.4.24-mtk/drivers/gpu/drm/mediatek/mtk_cec.c linux-4.4.24-mtk/drivers/gpu/drm/mediatek/mtk_cec.c