patch sonos_ubi_wl.patch
The default UBI behavior is to schedule the erases with the wear-level
worker thread.  What we have seen, in a very small number of cases, is that
if there is a power cut at the wrong point, some of the scheduled erases
have not yet taken place, and some of the update writes have, and the result
is an inconsistent_vid_hdr error, and a failure to mount any of the UBI
devices.  So we'll check for the updating condition and if we see it,
we're doing a VOLUP, and we'll do the erases synchronously.
diff -Naurp old/drivers/mtd/ubi/wl.c linux-4.4.24-mtk/drivers/mtd/ubi/wl.c