patch mtk_hdmi_cec_tx_arbitration.patch
This patch was provided by MTK to fix a suspected HW flaw in the MT8521p CEC peripheral.
Just before the peripheral attempts to transmit a frame it appears to be blind
to other potential traffic.  So if another device attempts to transmit during this
blind period there will be a collision on the bus as the two tranmissions will
occur during the same period.  On the wire this looks like two start bits in a row
followed by a single byte from the MT8521p.  The first start bit is from the other
device attempting to transmit.

This patch works around the issue by sampling the CEC line with a GPIO several
times before allowing the peripheral to actually transmit the frame.  This does
make the peripheral more prone to fail transmits because it cannot find a clear
window during heavy bus traffic but this is preferable to collisions.

The GPIO abitration is enabled via the gpio-base key on the cec DTB entry.
If the key is not present the feature is disabled and arbitration is fully
dependent on the HW.

This patch also includes minor improvements to logging which permits tracking
the lifecycle of a pending transmit.

diff -aurp -x .unpacked -x '*.orig' linux-4.4.24-mtk-orig/linux-4.4.24-mtk/drivers/gpu/drm/mediatek/mtk_cec.c linux-4.4.24-mtk/drivers/gpu/drm/mediatek/mtk_cec.c