patch sonos_hdmi_cec_receive_enonet_fix.patch
This patch removes a race bug in the CEC_RECEIVE ioctl which prevents reading data
that was advertised as available during poll().

If there is pending data on the transmit_queue when a device deconfigured, the CEC kthread
transmits cancels them.  Canceled data is moved to the msgs queue for receive which triggers
poll() to flag POLL_IN.  However if the CEC_RECEIVE ioctl is read -ENONET is returned
because the device is not configured.  This leaves the userspace program stuck
with a read event it cannot satisfy.

The fix is to allow reading while the device is deconfigured.  At worst there will
be no data and -EAGAIN will be returned.  Since the CEC kthread moves the pending
transmits to the receive queue rather than silently discarding them indicates they
are intended to be read so this appears the correct course of action.

diff -aurp -x .unpacked -x '*.orig' linux-4.4.24-mtk-orig/linux-4.4.24-mtk/drivers/media/cec/cec-api.c linux-4.4.24-mtk/drivers/media/cec/cec-api.c