patch sonos_mtk_disable_conn_pd.patch
At least two separate regions of kernel TEXT space are being corrupted on boot.
The first section is at 0xC0080400-0xC0080420 (32 bytes) and the second is 0xC008001C-0xC0080028
(12 bytes).  This corruption has probably been occurring for the entire duration
of the project however many factors of the kernel build determine whether the
corruption affects something critical.  Most builds the corruption goes unnoticed,
other builds crash or hang.

The specific mechanism of the corruption is not fully known.  However it was isolated
to the "CONN" Power Domain of the MT8521p.  This power domain oversees a proprietary
interconnect between the MT8521p and the MT6625 (an advanced power management IC).  Elrey
and Hideout do not use this PMIC so the interconnect isn't used.

The mechanism of the corruption is still being investigated but debugging has found that
the corruption is not related to any SW agent since protecting kernel memory (CONFIG_DEBUG_RODATA)
has no affect on the issue.  Therefore it is believed that the corruption is due to some HW action.
It is likely that enabling the power domain instigates some sort of DMA or memory mapping of the
non-existent MT6625 device into system RAM.  MTK has suggested this is likely as well and has provided
us with some more debugging tips.

This fix is to simply not allow the "CONN" power domain to power on.  The mtk-scpsys does not have
a mechanism for disabling the power domains from defconfig or DTB.  The power domain table is hardcoded
in the driver.  This changelist contains a patch (sonos_mtk_disable_conn_pd.patch) which adds a disable
field to the table and hard codes the CONN entry to disabled if CONFIG_SONOS is set.

REF: SWPBL-103467

diff -aurp -x .unpacked -x '*.orig' linux-4.4.24-mtk-orig/linux-4.4.24-mtk/drivers/soc/mediatek/mtk-scpsys.c linux-4.4.24-mtk/drivers/soc/mediatek/mtk-scpsys.c