patch 043_sonos_adv7511_fix_adv7533_support.patch
This patch updates the Linux ADV7511/33 driver to properly support the ADV7533 and ADV7535
MIPI bridge chips.  While the ADV7533 and ADV7535 are, according to Analog Devices, compatible
there are fairly large register differences between the ADV7511 and ADV7533/35 which are not
entirely accounted for in the driver.

This list of changes are as follows:
* Added explicit DTB compatibility for the ADV7535.  It simply aliases ADV7533 functionality.

* The regmap cache defaults for the ADV7511 are incorrect for the ADV7533/35.  I have added
an alternate default table for the ADV7533/35.

* ADV7533/35 initialization patches were missing a write of 0x10 to register 0x25, per
documentation.  The documentation does not explain what this or any of the other init patches do.

* Made the ADV7511_REG_CEC_TX_ENABLE and ADV7511_REG_CEC_RX_ENABLE registers volatile.  If
the cache gets out of sync then it's likely to cause problems.

* CEC Tx Retry configuration was split from one register field on ADV7511 to two register fields
on ADV7533/35.  One of the fields is in the same position as the 11 but is now specifically for
Low Error retries.  NACK retries shares a register with the Tx Enable bit which I combined into
a single write; executing separate field updates caused Tx to stop working entirely.

* The HPD control accessed by ADV7511_REG_POWER2 has become as single bit HPD override on the
ADV7533/35.

* Audio select accessed by ADV7511_REG_AUDIO_SOURCE has become a single bit field on ADV7533/35.
The values are the same, the ADV7511 just has more options available.

diff -aurp -x '*.orig' linux-4.9.99-orig/drivers/gpu/drm/bridge/adv7511/adv7511_audio.c linux-4.9.99/drivers/gpu/drm/bridge/adv7511/adv7511_audio.c