patch 050_sonos_adv7511_read_large_edids_fix.patch
This patch fixes an issue where the ADV7511 is unable to read EDIDs larger
than 256 bytes in size.  The stock ADV7511 driver in linux has a two issues
which resulted in the read EDID having the same 256 bytes reapeated for the
length of the true EDID.  This resulted in a checksum failure and the
rejection of the entire EDID.

The first issue is a simple unit conversion problem.  The ADV7511 reads the
EDID in 256 byte segments whereas Linux reads in 128 byte blocks.  In most
places the appropriate conversion was being accomplished but not at the
actual register write to change the EDID Segment.

The second issue is far more complicated.  The adv7511_get_edid_block()
function was written in a manner that is not consistent with the ADV7511/33/35
documentation.  If it needed to move the segment it would get the EDID controller
status and then only move the segment if the controller wasn't IDLE (status != 2).
It then proceeded to read the EDID from the controller regardless of whether the
segment was changed or not.  I really don't understand the intent here but it
does not work.

The ADV7511/33/35 reads the EDID automatically in two circumstances (that I know of):
	* HPD change from low to hi
	* EDID Segment is changed to a new value

With that in mind I have changed the EDID logic to the following:
	1. If the EDID controller is in reset (status == 0) error out because there's
	   nothing to read.
	2. If the EDID is currently being read (status == 1), wait for it to complete.
	3. Always retrieve the current segment from the HW to see if it actually needs
	   to be moved.  The EDID read will not occur if the segment doesn't change.
	4. If the segment needs to be changed, then change it and then wait for the
	   read to complete.

The other EDID controller status values involve HDCP which Apollo and the ADV7511
driver do not support so I have ignored them.

diff -aurp -x '*.orig' linux-4.9.99-orig/drivers/gpu/drm/bridge/adv7511/adv7511_drv.c linux-4.9.99/drivers/gpu/drm/bridge/adv7511/adv7511_drv.c