patch 055_sonos_simple_card_provide_clk_id.patch
This patch adds an optional parameter to the simple-card CPU/Codec DTB
subnodes so that the clock id can be specified.  The driver previously only
used clk_id 0 when configuring a CPU or Codec's set_sysclk callback.  For
most ALSA drivers this is fine however the fsl-esai-dai driver used the
clk_id to select which source clock to use for input/output.  A clk_id of 0
forced the ESAI to always use the system clock which cannot be divided down
to our target frequency of 44.1kHz.  Instead we want to use a clk_id of 1
which selects the Transmit EXTAL clock.  By specifying "clock-id" in the
simple-card DTB the correct clk_id can now be selected.  If no "clock-id"
is present then the original default of 0 is used.

diff -aurp -x '*.orig' linux-4.9.99-orig/Documentation/devicetree/bindings/sound/simple-card.txt linux-4.9.99/Documentation/devicetree/bindings/sound/simple-card.txt