patch 909_sonos_print_time.patch
Some users call first without a buffer to determine the size of the buffer needed.  The string used in the non-buffer case needs to match the string in the buffered case, as failing to do so can result in buffer overruns and memory corruption.

diff -Naurp old/kernel/printk/printk.c linux-4.9.99/kernel/printk/printk.c