patch 919_sonos_aml_spdif_mixer_ctrl_fixes.patch
This changlist fixes a number of issues with the ALSA mixer controls for
AMLogic's SPDIF ALSA driver.

The fixes are as follows:
* Changes the ALSA mixer control for input sample rate to provide a flat
integer in Hz instead of an enum which requires userspace to perform more
work to interpret.
* Changes the ALSA mixer controls for SPDIF CSB to use the standard ALSA
IEC958 format.  AMLogic used a non-standard method to read the CSB 4 bytes
at a time which was costly to implement in userspace.

This changelist is a WIP and the following issues are still outstanding:
* When SPDIF input is unlocked the sample rate is reported incorrectly.

diff -aurp -x '*.orig' linux-4.9.99-orig/sound/soc/amlogic/auge/audio_utils.c linux-4.9.99/sound/soc/amlogic/auge/audio_utils.c