patch 924_3_onos_aml_spdif_out_csb_fix.patch
This patch fixes an issue where the SPDIF OUT driver sets the CSB to a repeating
sequence of 4 bytes.  This is a complete violation of the IEC 958 specification
for the format of the 24 byte CSB structure.  Based on the format of the
iec958_chsts structure the CSB the driver should only be setting the first
4 bytes and leaving the remaining 20 bytes as zero.

diff -aurp -x '*.orig' linux-4.9.99-orig/sound/soc/amlogic/auge/spdif_hw.c linux-4.9.99/sound/soc/amlogic/auge/spdif_hw.c