patch 924_2_sonos_aml_fix_spdif_lock.patch
This patch adds a mechanism to scan the SPDIF preamble to determine if the A/B subframe swap has happened.
Since there is no reliable way to detect a lock/unlock in the driver, we run this scan on every DMA completion. It simply checks the B subframe for the appropriate preamble, which the hardware represents as 0x60. This data is stored out-of-band in the SPDIF buffer, so userspace ignores it. We only check the B subframe because the hardware represents the "first A subframe in the block" preamble the same way as the B preamble, so checking both would occasionally yield false positives.
diff -aurp -x '*.orig' linux-4.9.99-orig/sound/soc/amlogic/auge/ddr_mngr.c linux-4.9.99/sound/soc/amlogic/auge/ddr_mngr.c