patch 934_sonos_aml_spdif_isr_fixes.patch
The ISR for the SPDIFIN driver blindly acts upon all set bits in the IRQ status
register which can be set regardless of whether the IRQ mask enables it or not.
This means that if the ISR executes for an enabled IRQ reason and a disabled IRQ
happens to have its bit set in the status register, the code for that IRQ reason
will still be executed as well.

This patch fixes the issue by having the ISR read the IRQ mask register
and mask out any status bits not explicitly enabled.

diff -aurp -x '*.orig' linux-4.9.99-orig/sound/soc/amlogic/auge/spdif.c linux-4.9.99/sound/soc/amlogic/auge/spdif.c