sonos gpt lba
The A113 is looking for uboot at LBA 1, and in some places the GPT code
is hard-wired, assuming an MBR at LBA 0 and GPT header at LBA 1.  This
changes allows us to instantiate the GPT header at a different block.
Using 0x3000, offset 6MB, to start with, for no inherent or important
reason but for consistentcy - that's where the UBI partition was on
other platforms. MT8518S uses 0x5000, offset 10M for GPT due to
TEE/UBOOT