sonos disable console printk
Patch to add support for the enable_console bootarg in the serial core.

For transmit the write data will never reach the underlying driver.  The
characters are dropped in uart_write() and return a 100% completed write.
One side effect is uart_write() will start the UART HW transmitter but as
soon as the IRQ fires the underlying driver will stop the TX again when it
sees the buffer is empty (as if it had finished draining it).

For receive uart_insert_char() drops the characters sent from the underlying
driver on receive.