sonos aml spdif disable papb detection
This patch fixes an issue with the A113 SPDIFIN driver where the driver was
performing some sort of reset when a Dolby PaPb was detected in an unexpected location.
This can occur legally for a variety of reasons.  The fix from AML is to disable all
SPDIFIN IRQs except "find sample mode changed".  The following IRQs are now disabled:
    * valid changed
    * find nonpcm to pcm
    * find Pd changed
    * find Pc changed
    * find parity error
    * find overflow

Additionally this patch disables the entire "Find PaPb" function of the peripheral
which is controlled by bit 25 of the CTRL0 register.