Respect CONFIG_KSYMS_ALL when CONFIG_TRACEPOINTS set
An Amlogic-provided change that bypasses checking of
CONFIG_KALLSYMS_ALL to save RAM is generally harmless,
except in the case where LTTng is active, in which case
access via kallsyms is required to the symbols for certain
data structures.

Therefore, rework the logic to respect the value of
CONFIG_KSYMS_ALL if CONFIG_TRACEPOINTS is defined.