Fix S767 eARC Layout A channel count
ALSA does not support sample rates of >192kHz so high bit rate
encoded eARC data streams (Layout A) need to be reported as 8 channel
so the sample rate can be divided down properly.

This patch skips detection of the encoded data burst type based on the
PcPd value.  Instead encoded data streams are categorized as Layout A or
Layout B, which is then used to derive channel count. The new layout A/B
enums are only used in the eARC RX path to avoid breaking eARC TX (which we
don't use).

The PcPd detection feature of the peripheral is still enabled, it is just
ignored.

Layout B currently maps to a channel count of 8 because we do not have
a way to strip out the padding in HW yet.

Also disabled all debug prints in iec_info.c because they are very
spammy.