fix stream misalign issue
WARNING: THIS PATCH IS BROKEN!!

This was AML's first attempt to fix the S767 eARC Rx misalignment
issue.  It does not work in a way that is compatible with Sonos'
application but it must be included because it contains changes
and context which the real fix
(0002-AML-preamble-fix-for-earc-misalignment.patch) requires to
apply and function.