preamble for misalign issue
AMLogic's fix for the eARC misalignment issue.

The basic approach to this fix is to expand the DMA word size to include
the SPDIF preamble portion of the sammple.  The during the DMA completion
interrupt validate the Z sub-frame appears at the correct buffer offset.
If it is not it performs a fast reset of the audio device which hopefully
corrects the misalignment.

This requires 0001-BROKEN-AML-s-misalignment-fix.patch to apply and function.