Add phy_addr change support to CEC driver
This patch adds support to the S767 CEC driver for handling Physical Address
updates from the HDMI Tx driver.

It also removes a check from the CEC transmit function which performed a
HPD and physical address validity check if you attempted to transmit a
Report Physical Adress CEC message.  This is unnecessary validation
because the CEC message already has a physical address populated in it
and the check can cause a deadlock due to the ordering of mutexes taken
the transmit thread and the EDID read thread (when it updates the physical
address).

Finally it removes some code from aml_aocec_probe() which sets the CEC Physical
Address.  This looks like it was leftover from development testing because it
is unlikely that an EDID would have been read during driver probe as it occurs
fairly early in boot.