fix loses signal transitioning from 32K PCM to Dolby MAT
This AML patch fixes an issue where the application would get no eARC signal
occasionally when the source changes from 32k PCM to a Dolby MAT format.

The patch creates a timer which starts on device trigger and repeats every 1
second until the device is closed.  When the timer executes it checks the
current DMA pointer position and the state of the current PLL clock.  If the
clock indicates we should have a signal but the DMA pointer has not moved it
performs a PLL refresh which presumably kicks the peripheral into properly
parsing the eARC signal.

Change-Id: Iae32bf58e18a24f8a1edcec3e9c67e53521e79f0