Update eth speed unknown to zero
SPEED_UNKNOWN (-1) denoted an unspecified speed for a network interface, often equating to 65535.
This patch introduces a check during the conversion of link ksettings to legacy settings within
the ioctl request. If the speed is negative, implying an unknown state according to SPEED_UNKNOWN's
definition, the link_speed variable is now explicitly set to defined unknown_speed i.e. 0.

This change clarifies situations where the link speed is unknown, aiding userspace applications in
determining Ethernet port status, especially when the link is inactive or down.