Minor fixes to eARC Rx latency
This patch contains Sonos some minor changes for the 0001-audio-add-earcrx-latency-function.patch
provided by AMLogic.

First it changes the default eARC Rx Latency to 0 which is the value expected by
the Sonos application.

Second it reverts the changes to earcrx_get_latency() by AML's patch.
The original function returned the value of eARC Rx Latency in the peripheral
HW registers.  AML changed that behavior to return the value in the cached
latency variable.  The original behavior felt more accurate (but likely no
different in practice).