Use IRQ counter instead of DMA pointer for earc stall detection
See also: SWPBL-220127
The S767 eARC peripheral will occasionally stall. AML added a timer to
attempt to detect the stall and reset. However, this timer used the
current buffer position to see if the peripheral stalled. Since the
buffer wraps frequently, there is a chance that the timer will fire on a
wrap and will see the same position twice, which would be interpreted as
a stall and cause an unnecessary reset. This patch replaces the pointer
usage with a counter that gets incremented in every ISR. The counter
will take substantially longer to wrap so there's no longer a risk of a
false positive.