Adjust eARC buffer size for 192-sample periods
AML's eARC driver enforces a buffer size that is a multiple of the SPDIF
block size (192). If we use a period size of 192 and keep our ring size
at 128 periods, the worst-case buffer size will be larger than the
preallocated space by default. This patch sets the default preallocation
to its maximum to allow for this configuration.