Constrain eARC driver to integer periods
This is a follow up patch to 0001-eARC-round-pointer-to-period-size.patch.  That patch
has an implicit requirement that the number of periods in the buffer be an integer number.
Without the constraint asking for a 128 sample period results in a partial period at the
end of the buffer which causes the pointer tracking to become skewed everytime the buffer
wraps.