Prevent eARC parity errors from deferring timer
Parity errors cause a timer to be scheduled 10msec in the future which will reset
the PLL and hopefully resolve the parity error.  However if the timer is already
scheduled another parity error received before the 10msec is up will just defer the
timer for another 10msec.  It is possible and probable for the timer to be deferred
indefinitely by a stream of parity errors resulting in the curative reset never
occurring.  This patch changes that logic to only schedule a timer if one has not
already been scheduled.