improve songle flow control
This patch removes the callback-based 'enabled' parameter used to enable/disable
the HDMI ARC based on Songle values. Instead, it introduces the
'hdmitx_ext_set_songle_output' function, which is exported for use by the Sonos
hdmictl driver to directly control the Songle output. This change reduces the delay
overhead and establishes direct control over the Songle output. With this change, we
will be able to enable/disable the Songle output directly. Previously, we used to only
enable it due to ADC drift during HDMI is connected, but now we don't need to worry
about it and have control (enable/disable) on it based on Songle ADC values purely.