add rgb 8bit color attr list
Fix for Test ID 7-23
The reason for this change is that even though the CDF declares YCBCR = N,
the DUT can still output the YCBCR pixel format. This occurs because, in
the mesh_hdmi DRM driver, the `color_attr_list`, which defines the `color_space`
and `color_depth`, is selected based on the `crtc_eotf_type`, which always
returns YUV444 might be due to their HDMI IP design. This is then mapped
to the YCBCR format.

For HDMI TX compliance tests, we only need to ensure passing the RGB format
as per the CDF declaration. Therefore, We need to set RGB 8-bit as the
default format in the driver itself. According to our CDF, we only need to
pass Test ID 7-23, which requires RGB pixel encoding support as per the
specification stating, "All HDMI sources and sinks should support RGB 4:4:4
pixel encoding." This change will not affect any other tests we run;
it will just correct the issue for Test ID 7-23.