earc: Optimize zeroing with clear_user()
The `tmp_buf` is essentially a large zero buffer used to provide userspace with silence
when the DMA buffers are filled with unplayable garbage. We frequently call `earc_hw_params()`
whenever there is an audio format change, as we adjust configuration parameters of the
input device, such as sample rate and channel count. These frequent calls trigger the
use of `kmalloc` to allocate a large chunk of memory when earc is enabled or as audio starts,
which is released with earc disabled of invalid signal or as audio stops. Frequent `kmalloc` and
`kfree` calls can fragment memory over time, leading to reduced efficiency in the slab allocator,
especially if the system does not promptly reclaim and consolidate freed memory. While kfree releases
the memory, a low-priority kernel task is required to reclaim and consolidate the freed blocks
back into the slab

Removed the 'tmp_buf' a zero buffer to eliminate the need for frequent kmalloc and kfree operations.
Replaced the 'tmp_buf' with a single call to the clear_user() which is an already available function,
which directly zeros out the user-space buffer and eliminating unnecessary memory allocations.This
makes it easier to maintain and reducing memory overhead.