increase eARC HB Lost timeout to 170ms
After extensive testing on the Lost eARC Heartbeat (HB) issue, we have identified
the root cause. Occasionally, the TV sends a single shorter heartbeat signal,
consisting of only 8 pairs, with a shorter time interval, instead of the expected 10 pairs.

The current default eARC_LOST_HEARTBEAT timeout is set to 128ms, and as per the HDMI 2.1 spec,
the timeout should be between 120 and 130 ms. Between the Lost HB and the timeout, we did
receive a normal Good HB, but we never considered it to reset the eARC LostHB timeout counter,
causing a timeout and entry into the RX DISC 1 state, sending the comma sequence

Also noting that the heartbeat cadence is between 45 to 55 ms, in my testing with the TV, it is
45ms, meaning before the timeout (starting from the last received good heartbeat to the timeout),
there should be 2 heartbeats to assess whether we need to timeout or not. In this case, we received
one Lost HB signal followed by a normal Good HB, at which point we should have reset the eARC LostHB
timeout counter. However, this reset never occurred.

From what I understand, after the Lost HB, the s767 is considering the second Good HB signal after the
Lost HB. For example: Good HB → Lost HB → Good HB (not recognized) → Good HB (recognized and resets
the eARC timeout counter). To reach the second HB signal after the Lost HB signal, starting from the
last received heartbeat, it takes 45+45+45 = 135ms, which is the minimum heartbeat cadence. If the
heartbeat cadence is 55ms for maximum, then it would be 165ms.

AML, is that they are not considering subsequent HB signal after a Lost HB signal to reset the eARC
Lost HB timeout counter. Instead, they are waiting for one more normal HB signal to arrive before
resetting the counter. This reset flow is part of their existing RISC design, which they confirmed
is impossible to change. The only way to resolve this issue is by increasing the timeout in their
heartbeat lost timing register to allow time for one more normal HB signal to arrive before the
counter is reset. Looks like they haven't followed the HDMI 2.1 spec for the timeout consideration
for any reason.

This change is to increase the timoeut from 128ms to 170ms, AML's recommendation is to go for 170ms
as appropriate to fix this and I found it be valid, but as ran compliance tests again with this
change it failed HFR 5-2-25 test which is an HFR5-2-25: eARC RX gets Heartbeat Disconnect as known.
It was confirmed that this solution will have no side effects on the functionality itself, except
for the timeout compliance test.Here is the ticket for further followup on this issue SWPBL-237093.