eARC Add support to read eARC TX latency announced to eARC RX
This patch introduces a change to read the eARC TX latency value that is
announced to the eARC RX.

The ERX_LATENCY_REQ_REG register is read whenever an interrupt is detected
on 'INT_EARCRX_CMDC_STATUS_CH'.The value is then compared with the previously
stored latency value. If the current value differs, an alsactl event is
triggered to notify userspace about the change and readback the latency value.